New rcpp_ccl function There are actually four changes:
## ccl
is the same function as before, just w/o global variables

## ccl_old
the unchanged ccl function

## rcpp_ccl
a new function, written with Rcpp. It uses a different algorithm than the old ccl. A consequence is that the patch labels are not named identical because the sequence of patch identification is different (i.e. the patch w label 3 is not neccessarily the same patch in both algorithms) Otherwise the Algorithms seem comparable.

### rcpp_ccl only works with int matrices
it is sensitive to the type of matrix, therefore we need to change filter_matrix[...] <- 1 to filter_matrix[...] <- 1**L**